+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.0.0
+
+* Added example folder with a full Flutter demo project.
+
+---
+
+## 0.0.2
+* Added extensions for almost all types of data types.
+
+---
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release. No changes made.